deployment: Add a script to preserve consistent timestamps
..for files that did not change between deployments.
That is done by setting the modification time to the commit time.

That value is not a 100% true modification time value
(as the actual modification time is the time the file was last
changed on the server), but that'd be too difficult.
And the commit time should suffice and is at least deterministic..