Share generated resources between releases
..to ensure they are not suddenly unavailable
due tue a new deployment.
This is to prevent unavailable resources which
happen when the request that generated the resource (and links it)
happens before deployment and the actual download after deployment.

Note: we do not want to share typo3temp entirely as we never
want the code cache of one release to affect the code cache
of another.