Prepare to store secondary extensions in packages/
As composer repository paths allow php style globbing [1]
we can use one repository definition whith a wildcard (*)
to prepare for all local packages/extensions.

So that one only needs to add "foo_extension": "@dev" to
the "require" section.

Note that we still use /site/ as the path that contains
the extension which defines the site template, typoscript etc..
It's easier to type (site/ vs packages/your_site_slug/) and it clearly
documents that this is _the_ primary folder.

[1] https://getcomposer.org/doc/05-repositories.md#path