Enable UTF-8 filesystem by default, and use C.UTF-8 as generic locale
de_DE.UTF-8 (or any other language specifc locale) is not required
as LC_CTYPE conversion (which systemLocale is used for in TYPO3)
is requal for all UTF-8 locales.
Let's not pretend we set something like a default language here.

As we can use C.UTF-8 as generic locale it's fine to preset this
in AdditionalConfiguration.php and _not_ have require users
to configure this after installation.