composer.json: Move the site-package require statement up
Just a cosmetic thing. The point is:
When dependencies are required,
they will be added to the end of the list.
In that case the site-package will be somewhere hidden
between typo3/cms and other dependencies.
It's better to have it in the top to indicate:
That is the prime package.